[Codegen] Always use ? for non-zero offsets (#19952)
In order to rewrite subspans to buffer descriptors, we might need to be
able to fold offsets into the buffer descriptors. This means that we
need to be able to replace an offset with a different one (specifically
0) because the offset will be applied to the base pointer during buffer
casts. If the offset were dynamic, we can always memref.cast the
dynamic-ness of the offset back in, but we can't replace a static offset
with a different static offset. Therefore, never create buffers that
have a static non-zero offset during bufferization.