[GPU] Set insertion point to last slice index operand in reshape and slice fusion (#19959)
Empty tensor elimination relies on dominance of SSA values when
attempting to reuse buffers for slices of init operands. Ideally, empty
tensor elimination should be able to handle this, but it is difficult to
fix at that level. For now, this PR tries to avoid creating these
dominance issues in the first place.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>