[Codegen] Add support to emulate unsupported float type (#19943)
This change enables the conversion of types such as f8E4M3FNUZ and
f8E5M2FNUZ (emulated via the existing APIs) into f32 operations. The
conversion logic is now tightly coupled with the executable target
attribute, so that it is applied only for the gfx942 target. This
removes the need for manual pass configuration to specify source types
and aligns the behaviour with the target’s capabilities. For any new
conversion, just populate the conversion target with source and target
types.

FIX:
https://github.com/iree-org/iree/issues/19921#issuecomment-2643189839

---------

Signed-off-by: Chi Liu<Amos_Lewis@foxmail.com>
Co-authored-by: AmosLewis <Amos_Lewis@foxmail.com>