[GPU] Take element type bitwidth into account with vector size. (#19987)
This is for TileAndFuse workgroup and thread distribution logic. Before
this PR we only tried to vectorize to size 4 but now we vectorize to a
bitWidth of 128. This can however result in no vectorization for small
dimensions so the PR also introduces a fallback for smaller shapes to
vector size of 4.

Co-authored-by: MaheshRavishankar <mravisha@amd.com>
Signed-off-by: MaheshRavishankar <mravisha@amd.com>
Signed-off-by: Nirvedh <nirvedh@gmail.com>

---------

Signed-off-by: MaheshRavishankar <mravisha@amd.com>
Signed-off-by: Nirvedh <nirvedh@gmail.com>
Co-authored-by: MaheshRavishankar <mravisha@amd.com>