[DispatchCreation] Modify the generated fused op to not use concats. (#19980)
This is an almost complete rewrite of the pass to fuse contractions
horizontally which instead of concatenating operands to map to a GEMM,
followed by slices to extract the individual matmul results; the pass
now just creates a new operation with the operands being the common LHS,
the RHS of each of the gemms, and the output of each of the gemms. The
generated op yields the result of each constituent matmul.
This also allows for the RHS/output indexing maps of the gemms to be
mismatched, since only the LHS operand and indexing maps need to match.
The change also permutes the iteration space of the gemms to ensure that
the same indexing maps are used for the LHS across all the fused
matmuls.

The rest of the compiler stack has already been fixed up to handle such
operations.

---------

Signed-off-by: MaheshRavishankar <mravisha@amd.com>