Reland "[runtime][python] Allow to deref from VmRef to HalCommandBuffer" (#20058)
Re-lands PR#19629 that was previously reverted because one of the the
regression tests failed.

In IREE runtime, if we have a `iree.runtime.VmRef` of
`iree_hal_command_buffer_t` in python binding, it's currently not
possible to dereference it (via `VmRef.deref()`) into a
`iree.runtime.HalCommandBuffer` since the `VmRef::BindRefProtocol` is
missing for `HalCommandBuffer`.

In this PR it's added.

---------

Signed-off-by: PragmaTwice <twice@apache.org>
Co-authored-by: Twice <twice@apache.org>