[AMDGPU] Add RDNA4 targets and WMMA (#20048)
This PR adds definitions for the new WMMAs in RDNA 4, adds a target
definition, and updates tetsts.

The WMMA instructions have changed semantics (half as many inputs per
lane, no more weird broadcast thing and a different output) from RDNA3
so we now have WMMAR3 and WMMAR4 variants (the R4 are 0x19** in the
enum). Some tests are duplicated, some aren't yet. This is an initial PR
to enable RDNA4 bringup.