[ROCM] Don't reject RDNA4 targets, add them to wavesize32 list (#20098)
We have a check for not senting unknown targets
into LLVM that was causing silent failures in executable serialization.
This patch adds gfx120x as supported and also makes it pass +wavesize32
like gfx11 and gfx10.

This also fixes a TODO about just using amdgpu::Chipset to parse the
chipset version because I'm here already *and* fixes a bug in the device
global setup that would cause the device libraries to think we're in
wave64 mode when we're in wave32 mode