Increase size of the PotentialValuesState. (#20119)
When running the ElideTimepoints pass on a large program (llama 8b for
example) we exhaust our state, and fail to elide timepoints.

This significantly reduces the number of timepoints that end up in the
waiting set in those cases.

The number of timepoint waits in the stream output are as follows.
```
23765 << State Size 256
8801 << State Size 8192
```

Signed-off-by: Andrew Woloszyn <awoloszy@amd.com>