[DispatchCreation] Address missed horizontal fusion case with transposes  (#20060)
The check for finding valid gemms to fuse with did not permute the
indices of the LHS if required and use the permuted op to make the
decision. This change fixes this.

This change also actually triggers the horizontal fusions that were
missed earlier. For mi300 this keeps the existing performance, but
there is a slight slowdown for other AMD cards that we arent really
tracking at bleeding edge of performance. So adjust all the dispatch
counts and runtimes as seen with this PR.

Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>

---------

Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>