[GPU] Add support for conv padding when decomposing im2col (#20203)
Adds support for fused pad producer to linalg conv ops going down the
IGEMM path. This is done by swapping the pad with the extract slice of
the decomposed im2col op.
See this issue for details on flags needed to fuse the pad with the conv
in a single dispatch.
https://github.com/iree-org/iree/issues/20200

---------

Signed-off-by: Nirvedh Meshram <nirvedh@gmail.com>