[LLVMCPU] Set stack allocation limit as a HAL target property (#20209)
As long as the limit is defined arbitrarily as a pass-specific CLI
option, tiling/bufferization passes have no way of accounting for it.
This change would be the first step towards addressing compilation
failures in `LLVMCPUCheckIRBeforeLLVMConversionPass` systemically before
further tuning of the tiling/fusion logic gets done for particular
benchmarks.

A basic LIT test is also added to make sure that the limit can still be
overriden through a CLI option. This can be expanded further to test
other default attribute values.

Signed-off-by: Artem Gindinson <gindinson@roofline.ai>

---------

Signed-off-by: Artem Gindinson <gindinson@roofline.ai>