[LLVMGPU][DT] Add MaterializeDeviceEncodingPass to LLVMGPU passes behind flag (#19849)
This PR adds the `MaterializeDeviceEncodingPass` to LLVMGPU/Passes.cpp
behind a new flag `iree-llvmgpu-experimental-data-tiling`. The flag's
default value is false, because the codegen for data tiling ops on GPU
is not yet working and performant for all cases. Some of the work is in
flight, but it will likely take some time before data tiling codegen is
ready to be flipped on by default. For now, the flag allows developers
to enable the late materialization codegen path on LLVMGPU.

To effectively use the late materialization path for data-tiling fusion,
some additional non-default flags need to be set:
- `--iree-opt-data-tiling=false` (to turn off the early materialization
data tiling path)
- `--iree-dispatch-creation-experimental-data-tiling=true` (to turn on
the late materialization data tiling path)
- `--iree-dispatch-creation-pad-factor=128` (the current default is
based on CPU materialization)

This PR also includes a small fix to the ROCDLKernelConfig logic for
selecting the root op when there are pack/unpack ops in the dispatch.
The fix avoids selecting pack and unpack ops as root ops if possible.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>