[Runtime] Preserve denormals in floating point conversions (#20252)
These helper functions are used notably for printing and parsing, so
flushing denormals to zero there was not a good idea, even if actual
codegen would flush to zero. For example, flusing to zero when printing
could hide denormals, and parsing to zero could prevent reproducing
issues with denormals.

A couple of unrelated changes are lumped in this PR: dropping some
redundant branches to handle `nan_as_neg_zero` (which is used in FP8
types that have FN in their name). I just spotted that and it's
redundant because it's already handled at the end of the function, and
the tests still passing confirm that.

As rightly called out by @krzysz00 in
https://github.com/iree-org/iree/pull/20242#pullrequestreview-2686197346.

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>