[Preprocessing] Form single dispatch programs. (#20265)
The `MakeSingleDispatchForFunction` pass was an initial attempt to move
the entire program into a single dispatch. It was fairly basic and didnt
account for ABI etc. This PR changes the logic of the pass to find the
core computation of the function using slices and ignores ABI related
operations (and some additional ones).

The pass would previously handle multiple basic-blocks. We should be
able to handle them with the new approach, but needs some iteration on
how the upstream slicing utilities can be used in that context. Dropping
support for that for now, and will be revisited later. This pass is not
really used anywhere, so this "support" is not very meaningful.

---------

Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>