Make `math.fpowi` rewrite more systematic and test it. (#20293)
This is a follow-up to #20281 which lacked a test, and had left the c++
logic in a state that was prone to unintentional regression. Also, this
ensures that the `math.fpowi` rewrite happens even with
`-iree-codegen-gpu-native-math-precision`, which is intentional as
`math.fpowi` is not really a math function, like the comment explains.

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>