[Stream] Implement folders for stream.tensor.encode op. (#20298)
There are few folders available for `stream.tensor.encode` op, we can
fold it into the source resource for the cases:
- The source encoding is as the same as the result encoding.
- If they are not identical, fold the op when the source layout and the
result layout are identity layouts.

A tensor type without encodings is recognized as an identity layout. A
tensor type with the encoding is an identity layout if the encoding
attribute implements the `isIdentityLayout` interface method and it
returns true.

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>