[Stream] Add support for materializing stream.tensor.encode ops. (#20321)
The revision implements the MaterializeEncodeOps pass. The pass runs the
canonicalization patterns from `stream.tensor.encode` and folds NOP
`stream.tensor.encode` ops away. Then it creates executables and
replaces the `stream.tensor.encode` ops with `stream.async.dispatch` ops
on the executables and entries.

It caches the executables and entries based on the (source encoding,
result encoding) pairs. The encode ops that have the same encoding
signature can launch the same executables. The only difference is about
the dynamic dimension sizes, and they are passed as function arguments.
It means that the executable has enough context to deal with different
arguments (for dynamic sizes) in different dispatches.

The revision also adds a builder to `stream.async.dispatch` op that
takes an external export op.

It is a step towards https://github.com/iree-org/iree/issues/20187

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>