Avoid nested fashion traversal in ModuleOp when possible. (#20344)
There are three cases in the revision:

1. Gather `ExecutableVariantOp` ops. The assumption is that we always
have `module{executable{variant}}}` structrue. I.e., there are no nested
executables variants. In the context, we can always use getOps from the
module ops and executable ops to gather the variant ops.
2. Gather the HAL ops like `ExecutableConstantLoadOp` and
`DeviceMemoizeOp`. They are always in the host code function, so it
iterates over the top-level callables and walk on them.
3. Update global ops. The global ops are always the top-level ops in
modules.

Signed-off-by: hanhanW <hanhan0912@gmail.com>