[GPU] Add canonicalize and cse after im2col decomposition (#20414)
The reason to do this is that we have swap extract slice with pad
pattern after im2col decomposition in the pass and this can lead to
tensor.pad + tensor.cast which would get folded after canonicalization.
However, we run vectorization passes before canonicalizing and this will
cause tensor pad to not vectorize as the output dimensions will be
dynamic.

Signed-off-by: Nirvedh <nirvedh@gmail.com>