[GPU] Update the cache info for padding resolver in cloneWithSimplifiedConfig. (#20371)
The revision creates an empty GPUPadLayoutAttr in the first place, and
materialize the cache config in the interface method. The cache config
is only used in encoding specialization and materialization. We do not
need such information, which is sort of duplicated config, in any other
lowering. Thus, we do not need to hardcode the cache config in the first
place. They can be resolved by the new utility (i.e., `getL1CacheInfo`)
when we need it.

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>